+++ docs/changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 ~~~~~~~~~
 
+Unreleased
+----------
+
+* User can pass an `Options` instance (from `qiskit-ibm-runtime`)
+  via a keyword argument to both an `IBMQBackend` constructor and 
+  an instance method `IBMQBackend.process_circuits`.
+
 0.53.0 (April 2024)
 -------------------